Re-organizing enums + state definitions
